+++ Argos/SSHSessionManagement.swift
@@ -13,7 +13,7 @@
 //
 
 import Foundation
-import Citadel
+@preconcurrency import Citadel
 import NIOCore
 
 extension SSHService {